Add categories with html files

+++ src/main/java/org/launchcode/munchincrunch/data/CategoryRepository.java
@@ -3,5 +3,5 @@ package org.launchcode.munchincrunch.data;
 import org.launchcode.munchincrunch.models.Category;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Long> {
+public interface CategoryRepository extends CrudRepository<Category, Integer> {
 }
